// sourcecode/BTrak.Packages/snova-trading-module/trading-module-components/trading-components/src/lib/billing/models/userLevelAccess.ts
export class UserLevelAccess {
        id: any;
        userId: any;
        programId: any;
        levelIds: any;
        levelXml: any;
        fullName: any;
        levelId: any;
        clientId: any;
        roleIds: any;
        roleXml: any;
        referenceText: any;
        levelName: any;
}